+++ src/components/Project/Project.js
@@ -1,13 +1,18 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-const Project = ({id, name, tasksIds}) => {
+const Project = ({id, name, tasksIds, onClick}) => {
     const prjPath = `/projects/${id}`
+
+    const handleClick = (event) => {
+        onClick(id)
+    }
+
     return (
         <div>
             <p>{id}. Project name: {name}</p>
             <Link to={prjPath}>
-                <button>To Tasks</button>
+                <button onClick={handleClick}>To Tasks</button>
             </Link>
         </div>
     )
