+++ src/components/ProjectList/ProjectList.js
@@ -1,86 +1,14 @@
 import React from 'react'
+import { BrowserRouter, Switch, Route, Link, Redirect, withRouter, Router  } from 'react-router-dom';
+
 import NewProject from '../NewProject/NewProject'
 import Project from '../Project/Project'
-
-/*
-const DataProjectNotNorm = [
-    {
-      id: 0,
-      name: 'Project 1',
-      tasks: [
-        {
-          id: 0,
-          name: '1.1',
-          description: 'www',
-          completed: false
-        },
-        {
-            id: 1,
-            name: '1.2',
-            description: 'eee',
-            completed: true
-          },
-      ]
-    },
-    {
-        id: 1,
-        name: 'Project 2',
-        tasks: [
-          {
-            id: 0,
-            name: '2.1',
-            description: 'ggg',
-            completed: false
-          },
-          {
-              id: 1,
-              name: '2.2',
-              description: 'lhnv',
-              completed: false
-            },
-        ]
-    },
-]
-*/
-
-const dataproject = {
-    projectsById: {
-      1: {
-        id: 1,
-        name: 'Учеба',
-        tasksIds: [1]
-      },
-      2: {
-        id: 2,
-        name: 'Дом',
-        tasksIds: [2]
-      },
-      // ...
-    },
-  
-    tasksById: {
-      1: {
-        id: 1,
-        name: 'Task #1',
-        description: 'descr',
-        completed: false,
-      },
-      2: {
-        id: 2,
-        name: 'Task #2',
-        description: 'descr',
-        completed: true,
-      },
-      // ...
-    }
-  }
-
-const NormalisationState = (stateToNorm) => {
-    return stateToNorm
-}
+import dataproject from '../Data/Data'
+import StartPage from '../StartPage/StartPage'
+import TaskList from '../TaskList/TaskList'
 
 class ProjectList extends React.Component {
-    state = NormalisationState(dataproject)
+    state = dataproject
 
     newProj = (name) =>{
       this.setState(curSt =>{
@@ -93,14 +21,31 @@ class ProjectList extends React.Component {
         return {projectsById: newSt}
       })
     }
+    
+    //ProjIdPrint =
+
+    ProjectIdentify = (id) => {
+      const oldState = this.state
+      const newState = {...oldState, idProjRender: id  }
+      this.state = newState
+    }
 
     render() {
         return (
             <div>
-                <NewProject newProj={this.newProj}/>
-                {Object.values(this.state.projectsById).map(prj => 
-                    <Project id={prj.id} tasksIds={prj.tasksById} name={prj.name}/>
-                )}
+                <Switch>
+                  <Route exact path="/" component={StartPage}/>
+                  <Route exact path="/projects">
+                    <NewProject newProj={this.newProj}/>
+                    {Object.values(this.state.projectsById).map(prj => 
+                      <Project key={prj.id} tasksIds={prj.tasksById} 
+                        name={prj.name} onClick={ProjectIdentify}/>
+                    )}
+                  </Route>
+                  <Route exact path='/projects/:projId'>
+                    <TaskList state={this.state}/>
+                  </Route>
+                </Switch>
             </div>
         )
     }
